changed file extensions (ts -> mts, js -> mjs)
This fixes the issues with importing modules from the tests, although the tests still fail due to references to the DOM which does not exist in NodeJS.

I didn't change the file extenson on main.ts since it should never be imported by a test and also since if I change it to a .ts extension the Electron app throws an error on launch.

I also included some changes to debug-settings.ts that I forgot to commit a while ago.