replaced `Cursor.predecessor` with `Cursor.position`
The position is a number that tracks how far along the cursor is (it counts the number of MathComponents before the cursor, so 0 means the cursor is at the beginning).

The reason for this change was because I realized that a lot of the time you need the position more than the predecessor. (The predecessor can still be accessed via a `predecessor` method if I choose to write such a method).