rewrote Cursor.fromClick
This version has a different algorithm: it finds the group (MathComponentGroup or line) that you clicked on and then finds the component within that group that minimizes the horizontal distance to your cursor. This has different (better) behavior, for example when you start selecting something in the numerator of a fraction and drag your cursor all the way to the end of the line.