separated initialization logic
The app construction is now done in App.mts, while the call to the `initialize` method is in app-initializer.mts. Also, the initializers for individual math components are now in separate files, which fixes the issues that occurred when the components were loaded (and initialized) as a dependency in tests. The new module loading in the HTML file also ensures that any initialization gets done even if nothing else imports the module.

I also updated the debug settings file.