fixed selection behavior at ends of lines
Previously, if you selected and dragged all the way to the right, it would select the line break at the end. Now it does not.

Previously, if you selected and dragged all the way to the left, it would display your cursor at the end of the previous line (after the line break). Now it does not.