fixed incorrect key handlers
Some of the key handlers didn't have the `shiftKey` property set, which causes problems if they are using a character that can only be typed by pressing shift (e.g. an asterisk). Also, some of the handlers for inserting components didn't clear the autocomplete. Both of these issues are now fixed.