slightly modified cursor movement when text is selected
After this change, if your cursor is at the beginning of a selection and you press right, it will clear the selection and move to the end of the selection. Similarly, if your cursor is at the end of a selection and you press left, it will clear the selection and move to the beginning of the selection. (This is how it works in most other programs).