fixed incorrectly failing tests
The unit test for rendering App tabs was failing because I added a close button to the tabs and forgot to update the test to account for this.

The integration test was failing because I added tabs, which changed how the app looks, and I forgot to update the tests.

I have now added a precommit hook (not committed) that runs the unit tests before every commit, so these issues should occur less.