optimized partitionArray
When it is called by Cursor.fromClick, the elements are already in the right order and just need to be grouped into a list of lists. This allows for an optimization where you can skip checking the rest of the array one you've found one that isn't in the current group.