optimized MathSymbol.render: removed HTML parsing
Before the optimization, that one line of code (`span.innerHTML = this.symbol`) accounted for ~28% of the total time when pressing a key in the performance test document. Now it only takes ~5%.

However, somehow the total time for each keypress does not seem to have changed significantly (it still takes 150-250 milliseconds per keypress for the performance test document).