significantly optimized key handling when typing
This commit merged the branch `performance-optimizations` into `master`.

After the optimization, when pressing a key to type a character, it no longer re-renders the whole document. Instead, it cleverly inserts the new character into the document structure.

This commit also introduced the `LiveRenderer` class, which contains helper functions that handle the modification of the rendered document.